# Make histogram function faster