@opennota 2015-12-03
I found a way to speed it up even further. I'll update the PR soon.