@opennota 2015-12-03
benchstat:

```
name                       old time/op    new time/op    delta
EncodeDigitsConstant1e4-4    64.9µs ± 0%    63.5µs ± 0%  -2.17%  (p=0.008 n=5+5)
EncodeDigitsConstant1e5-4     610µs ± 0%     598µs ± 0%  -2.02%  (p=0.008 n=5+5)
EncodeDigitsConstant1e6-4    6.12ms ± 0%    6.00ms ± 0%  -1.97%  (p=0.008 n=5+5)
EncodeDigitsSpeed1e4-4        224µs ± 0%     225µs ± 0%    ~     (p=0.333 n=4+5)
EncodeDigitsSpeed1e5-4       2.11ms ± 0%    2.11ms ± 0%    ~     (p=0.421 n=5+5)
EncodeDigitsSpeed1e6-4       21.0ms ± 0%    21.0ms ± 1%    ~     (p=0.690 n=5+5)
EncodeDigitsDefault1e4-4      439µs ± 0%     439µs ± 0%    ~     (p=0.690 n=5+5)
EncodeDigitsDefault1e5-4     6.01ms ± 1%    6.00ms ± 1%    ~     (p=0.690 n=5+5)
EncodeDigitsDefault1e6-4     63.2ms ± 0%    63.5ms ± 2%    ~     (p=0.690 n=5+5)
EncodeDigitsCompress1e4-4     440µs ± 0%     439µs ± 0%    ~     (p=0.222 n=5+5)
EncodeDigitsCompress1e5-4    6.00ms ± 1%    6.00ms ± 2%    ~     (p=0.690 n=5+5)
EncodeDigitsCompress1e6-4    63.1ms ± 1%    63.1ms ± 1%    ~     (p=1.000 n=5+5)
EncodeTwainConstant1e4-4     93.0µs ± 1%    90.7µs ± 0%  -2.54%  (p=0.008 n=5+5)
EncodeTwainConstant1e5-4      701µs ± 0%     682µs ± 0%  -2.78%  (p=0.016 n=5+4)
EncodeTwainConstant1e6-4     6.96ms ± 0%    6.77ms ± 0%  -2.67%  (p=0.008 n=5+5)
EncodeTwainSpeed1e4-4         248µs ± 0%     248µs ± 0%    ~     (p=0.556 n=4+5)
EncodeTwainSpeed1e5-4        2.00ms ± 0%    2.00ms ± 0%    ~     (p=0.548 n=5+5)
EncodeTwainSpeed1e6-4        19.5ms ± 0%    19.5ms ± 1%    ~     (p=0.095 n=5+5)
EncodeTwainDefault1e4-4       491µs ± 0%     491µs ± 1%    ~     (p=0.690 n=5+5)
EncodeTwainDefault1e5-4      6.86ms ± 2%    6.82ms ± 1%    ~     (p=1.000 n=5+5)
EncodeTwainDefault1e6-4      71.7ms ± 1%    71.6ms ± 1%    ~     (p=0.690 n=5+5)
EncodeTwainCompress1e4-4      496µs ± 0%     496µs ± 1%    ~     (p=0.548 n=5+5)
EncodeTwainCompress1e5-4     7.55ms ± 3%    7.57ms ± 2%    ~     (p=0.690 n=5+5)
EncodeTwainCompress1e6-4     80.7ms ± 1%    81.3ms ± 3%    ~     (p=0.690 n=5+5)

name                       old speed      new speed      delta
EncodeDigitsConstant1e4-4   154MB/s ± 0%   158MB/s ± 0%  +2.21%  (p=0.008 n=5+5)
EncodeDigitsConstant1e5-4   164MB/s ± 0%   167MB/s ± 0%  +2.06%  (p=0.008 n=5+5)
EncodeDigitsConstant1e6-4   163MB/s ± 0%   167MB/s ± 0%  +2.01%  (p=0.008 n=5+5)
EncodeDigitsSpeed1e4-4     44.5MB/s ± 0%  44.5MB/s ± 0%    ~     (p=0.286 n=4+5)
EncodeDigitsSpeed1e5-4     47.3MB/s ± 0%  47.3MB/s ± 0%    ~     (p=0.508 n=5+5)
EncodeDigitsSpeed1e6-4     47.7MB/s ± 0%  47.6MB/s ± 1%    ~     (p=0.571 n=5+5)
EncodeDigitsDefault1e4-4   22.8MB/s ± 0%  22.8MB/s ± 0%    ~     (p=0.667 n=5+5)
EncodeDigitsDefault1e5-4   16.6MB/s ± 1%  16.7MB/s ± 1%    ~     (p=0.540 n=5+5)
EncodeDigitsDefault1e6-4   15.8MB/s ± 0%  15.7MB/s ± 2%    ~     (p=0.722 n=5+5)
EncodeDigitsCompress1e4-4  22.7MB/s ± 0%  22.8MB/s ± 0%    ~     (p=0.238 n=5+5)
EncodeDigitsCompress1e5-4  16.7MB/s ± 1%  16.7MB/s ± 2%    ~     (p=0.595 n=5+5)
EncodeDigitsCompress1e6-4  15.9MB/s ± 1%  15.9MB/s ± 1%    ~     (p=0.889 n=5+5)
EncodeTwainConstant1e4-4    108MB/s ± 1%   110MB/s ± 0%  +2.60%  (p=0.008 n=5+5)
EncodeTwainConstant1e5-4    143MB/s ± 0%   147MB/s ± 0%  +2.86%  (p=0.016 n=5+4)
EncodeTwainConstant1e6-4    144MB/s ± 0%   148MB/s ± 0%  +2.74%  (p=0.008 n=5+5)
EncodeTwainSpeed1e4-4      40.4MB/s ± 0%  40.4MB/s ± 0%    ~     (p=0.619 n=4+5)
EncodeTwainSpeed1e5-4      50.1MB/s ± 0%  50.0MB/s ± 0%    ~     (p=0.532 n=5+5)
EncodeTwainSpeed1e6-4      51.3MB/s ± 0%  51.2MB/s ± 1%    ~     (p=0.143 n=5+5)
EncodeTwainDefault1e4-4    20.4MB/s ± 0%  20.4MB/s ± 1%    ~     (p=0.706 n=5+5)
EncodeTwainDefault1e5-4    14.6MB/s ± 2%  14.7MB/s ± 1%    ~     (p=0.937 n=5+5)
EncodeTwainDefault1e6-4    13.9MB/s ± 1%  14.0MB/s ± 1%    ~     (p=0.690 n=5+5)
EncodeTwainCompress1e4-4   20.2MB/s ± 0%  20.2MB/s ± 1%    ~     (p=0.746 n=5+5)
EncodeTwainCompress1e5-4   13.3MB/s ± 3%  13.2MB/s ± 2%    ~     (p=0.579 n=5+5)
EncodeTwainCompress1e6-4   12.4MB/s ± 1%  12.3MB/s ± 2%    ~     (p=0.690 n=5+5)
```