@opennota 2015-12-06
Zeroing the histogram in assembler like this

```
    XORPS X0, X0
    MOVOU X0, (DI)
    MOVOU X0, 16(DI)
    MOVOU X0, 32(DI)
...
    MOVOU X0, 1008(DI)
```
seems to be 0.5-1% faster. Probably not enough to justify the change.